requirements.txt
flask
google-generativeai
python-dotenv
gunicorn
supabase
groq
Flask-Mail